requirements.txt
alembic==1.18.4
bleach==6.3.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.3
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gunicorn==23.0.0
greenlet==3.3.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
Pillow==11.3.0
paramiko==3.5.1
openpyxl==3.1.5
SQLAlchemy==2.0.48
tinycss2==1.4.0
typing_extensions==4.15.0
Werkzeug==3.1.6
WTForms==3.2.1
python-dotenv==1.1.1

# python -m venv .venv
# source .venv/bin/activate
# pip install -r requirements.txt
# sudo journalctl -u contable -n 30
# Para crear el servicio:
# sudo nano /etc/systemd/system/contable.service
# [Unit]
# Description=contable Service
# After=network.target  
# [Service]
# User=tu_usuario
# WorkingDirectory=/ruta/a/tu/proyecto
# ExecStart=/ruta/a/tu/proyecto/.venv/bin/gunicorn -w
# 4 -b 0.0.0.0:8000 app:app
# Restart=always
# [Install]
# WantedBy=multi-user.target

# sudo systemctl daemon-reload && sudo systemctl enable contable && sudo systemctl restart contable